Switched to echo framework
Now bear with me here

I realize I just made an argument for learning the standard library, but not only is the echo framework simplify my code considerably, the authentication framework i will be using has support for echo, but not the standard library.

Which is all fine, because at the end of the day the reason frameworks exist is because developer sat at their computer with the standard library and decided some things needed simplifying.

Learning this lesson for myself and consciously picking a framework is a good conclusion.